refactor schema and handlers
